Added password input
