+++ src/dashboard/plots.py
@@ -1,0 +1,223 @@
+"""plots.py - Create plots for the dashboard"""
+
+# Get packages.
+import altair as alt
+import pandas as pd
+import streamlit as st
+
+
+def format_data_for_table(raw_df: pd.DataFrame) -> pd.DataFrame:
+    """Format the data for display in Streamlit.
+    Args:
+        data_df (pd.DataFrame): The data to be formatted.
+
+    Returns:
+        pd.DataFrame: The formatted data.
+    """
+    # Group the data by the specified columns
+    grouped = raw_df.groupby(
+        ["Date", "Aircraft", "AircraftCommander", "SecondPilot", "Duty"]
+    )
+
+    # Aggregate to sum the FlightTime for each group
+    flight_time_sum = grouped.agg(
+        FlightTime=("FlightTime", "sum")
+    ).reset_index()
+
+    # Calculate the size of each group (number of launches)
+    group_sizes = grouped.size().reset_index(name='Launches')
+
+    # Merge the sum of FlightTime and the count of Launches for each group
+    data_df = pd.merge(
+        flight_time_sum,
+        group_sizes,
+        on=["Date", "Aircraft", "AircraftCommander", "SecondPilot", "Duty"]
+    )
+
+    # Sort by date in descending order.
+    data_df = data_df.sort_values(by="Date", ascending=False)
+
+    # Format the date.
+    data_df["Date"] = data_df["Date"].dt.strftime("%d %b %y")
+
+    # Convert the FlightTime (minutes) to a string in HH:MM format.
+    data_df["FlightTime"] = data_df["FlightTime"].apply(
+        lambda x: f"{x//60}:{x % 60:02d}"
+    )
+
+    # Reorder the columns
+    desired_order = ["Date", "Aircraft", "AircraftCommander",
+                     "SecondPilot", "Duty", "Launches", "FlightTime"]
+    # Ensure all desired columns are in the DataFrame before
+    # reordering to avoid KeyError. This also implicitly filters
+    # out any columns not listed in `desired_order`
+    data_df = data_df[desired_order]
+    return data_df
+
+
+def plot_launches_by_commander(df: pd.DataFrame):
+    """ Plot the number of launches by AircraftCommander.
+
+    Args:
+        df (pd.DataFrame): The data to be plotted.
+    """
+    # Group by AircraftCommander and count launches
+    launches_by_commander = df.groupby("AircraftCommander").size()
+    launches_by_commander = launches_by_commander.reset_index(
+        name='Number of Launches'
+    )
+
+    # Sort launches by commander in descending order explicitly in the chart
+    chart = alt.Chart(launches_by_commander).mark_bar().encode(
+        # Quantitative scale for Number of Launches
+        x=alt.X('Number of Launches:Q', sort=None),
+        # Sort by Number of Launches in descending order
+        y=alt.Y('AircraftCommander:N', sort='-x'),
+        color=alt.value('blue'),  # Set bar color
+        tooltip=['AircraftCommander', 'Number of Launches']  # Tooltip on hover
+    ).properties(
+        title='Launches by Aircraft Commander'
+    ).configure_axis(
+        labelFontSize=12,
+        titleFontSize=14
+    )
+
+    # Display the chart in Streamlit.
+    st.altair_chart(chart, use_container_width=True)
+
+
+def plot_all_launches(df: pd.DataFrame):
+    """ Plot all launches in the data.
+
+    Args:
+        df (pd.DataFrame): The data to be plotted.
+    """
+    # Sort the data by date in descending order.
+    df = df.sort_values(by="Date", ascending=False)
+    df = df.drop(columns=["_id"])
+
+    # Format the date.
+    df["Date"] = df["Date"].dt.strftime("%d %b %y")
+
+    # Convert the FlightTime (minutes) to a string in HH:MM format.
+    df["FlightTime"] = df["FlightTime"].apply(
+        lambda x: f"{x//60}:{x % 60:02d}"
+    )
+
+    # Format TakeOffTime and LandingTime.
+    df["TakeOffTime"] = df["TakeOffTime"].dt.strftime("%H:%M")
+    df["LandingTime"] = df["LandingTime"].dt.strftime("%H:%M")
+
+    # Make date the first column.
+    df = df[["Date"] + [col for col in df.columns if col != "Date"]]
+
+    # Plot all data.
+    st.dataframe(df, use_container_width=True)
+
+
+def show_logbook_helper(df: pd.DataFrame, commander: str):
+    """Show the number of launches by AircraftCommander in a table.
+
+    Args:
+        df (pd.DataFrame): The data to be displayed.
+        commander (str): The AircraftCommander to filter by.
+    """
+    # Filter the data by AircraftCommander, if specified.
+    if commander:
+        # Get launches where the pilot is commander.
+        filtered_df = df[df["AircraftCommander"] == commander]
+
+        # Get launches where the pilot is second pilot and the
+        # duty contains SCT or AGT.
+        second_pilot_df = df[df["SecondPilot"] == commander]
+        sct_df = second_pilot_df[second_pilot_df["Duty"].str.contains(
+            "SCT|AGT", case=False
+        )]
+
+        # Merge the commander and sct dataframes.
+        filtered_df = pd.concat([filtered_df, sct_df])
+
+        # Sort the data by date in descending order.
+        filtered_df = filtered_df.sort_values(by="Date", ascending=False)
+    else:
+        filtered_df = df
+        commander = "All"
+
+    # Format the data for display in Streamlit.
+    display_df = format_data_for_table(filtered_df)
+    st.header("Logbook Helper")
+    st.text(f"Launches by {commander}")
+    st.dataframe(
+        data=display_df,
+        hide_index=True,
+        use_container_width=True
+    )
+
+
+def quarterly_summary(df: pd.DataFrame,
+                      commander: str,
+                      quarter: str) -> pd.DataFrame:
+    """ Show a quarterly summary of the number of launches
+    for each AircraftCommander.
+
+    Args:
+        df (pd.DataFrame): The data to be summarized.
+        commander (str): The AircraftCommander to filter by.
+        quarter (str): The quarter to display."""
+
+    # Get all elements where the pilot is commander.
+    commander_df = df[df["AircraftCommander"] == commander]
+
+    # Get elements where the duty contains SCT or AGT and the pilot
+    # is second pilot.
+    sct_df = df[df["Duty"].str.contains(
+        "SCT|AGT", case=False
+    )]
+    sct_df = sct_df[sct_df["SecondPilot"] == commander]
+
+    # Merge the commander and sct dataframes.
+    commander_df = pd.concat([commander_df, sct_df])
+
+    # Extract the quarter from the date.
+    quarterly_df = commander_df.copy()
+    quarterly_df["Quarter"] = quarterly_df["Date"].dt.to_period("Q")
+    quarterly_df = quarterly_df[quarterly_df["Quarter"] == quarter]
+    quarterly_df = quarterly_df.drop(columns=["Quarter"])
+
+    # Find the last date where PLF was true. This is the last date where:
+    # - 'SecondPilot' is commander
+    # - 'PLF' is true
+    # - 'Duty' contains 'SCT'
+    # Find the last SCT and PLF dates.
+    sct_in_quarter = sct_df[
+        sct_df["Date"].dt.to_period("Q") <= quarter
+    ]
+
+    if sct_in_quarter.empty:
+        last_sct = "N/A"
+        last_plf = "N/A"
+    else:
+        last_sct = sct_in_quarter["Date"].max().strftime("%d %b %y")
+        last_plf = sct_in_quarter["Date"].max().strftime("%d %b %y")
+
+    # Create a summary table for the selected quarter.
+    # Count launches and hours flown by AircraftCommander.
+    summary = pd.DataFrame({
+        "Aircraft Commander": commander,
+        "Launches": quarterly_df.shape[0],
+        "Hours": quarterly_df["FlightTime"].sum(),
+        "Last SCT": last_sct,
+        "Last PLF": last_plf
+    }, index=[0])
+
+    # Convert the FlightTime (minutes) to a string in HH:MM format.
+    summary["Hours"] = summary["Hours"].apply(
+        lambda x: f"{x//60}:{x % 60:02d}"
+    )
+
+    # Display the summary table.
+    st.header("Quarterly Summary Helper")
+    st.dataframe(
+        data=summary,
+        hide_index=True,
+    )
